I hope this shit works...

+++ lib/database.cpp
@@ -3,7 +3,7 @@
 Database::Database()
 {
 #if DEBUG_THR
-    cout << "Database: olá" << endl;
+    cout << "Database: ola" << endl;
 #endif
 }
 
@@ -14,28 +14,34 @@ Database::~Database()
 #endif
 }
 
-void Database::start()
+int Database::start()
 {
-    status = -1;
-    while (!connectDatabase())
+    if (status < 0)
     {
-        cout << "Not connected to Database\n";
+        if (connectDatabase())
+            status = 1;
+        else
+            status = -1;
     }
+
+    return status;
 }
 
 void Database::stop()
 {
     // write db status = -1
-    PGresult *result;
-    std::string order;
-    CleanTable("mesorder");
-    if (PQstatus(dbconn) == CONNECTION_OK)
+    if (status > 0)
     {
-        order = "Update requests \n SET status = -1;";
-        result = PQexec(dbconn, order.c_str());
-        PQclear(result);
+        PGresult *result;
+        std::string order;
+        CleanTable("mesorder");
+        if (PQstatus(dbconn) == CONNECTION_OK)
+        {
+            order = "Update requests \n SET status = -1;";
+            result = PQexec(dbconn, order.c_str());
+            PQclear(result);
+        }
     }
-
 }
 
 int Database::connectDatabase()
@@ -56,39 +62,42 @@ int Database::connectDatabase()
 
 int Database::checkProgressWorking()
 {
-    PGresult *result;
-    std::string order;
-    int dbStatus; // get this value from database
-
-    if (PQstatus(dbconn) == CONNECTION_OK)
+    if (status > 0)
     {
-        order = "SELECT status FROM requests;";
-        result = PQexec(dbconn, order.c_str());
-        dbStatus = atoi(PQgetvalue(result, 0, 0));
-        PQclear(result);
-    }
-
-    if (status == -1 && dbStatus == 1)
-    { // previously working
-        return -1;
-    }
-    if (status == -1 && dbStatus == -1)
-    { // start working
-        status = 1;
-        // write status on database
         PGresult *result;
         std::string order;
+        int dbStatus; // get this value from database
+
         if (PQstatus(dbconn) == CONNECTION_OK)
         {
-            order = "Update requests \n SET status = 1;";
+            order = "SELECT status FROM requests;";
             result = PQexec(dbconn, order.c_str());
+            dbStatus = atoi(PQgetvalue(result, 0, 0));
             PQclear(result);
         }
-        return 1;
-    }
-    if (status && dbStatus)
-    { // working with connection
-        return 1;
+
+        if (status == -1 && dbStatus == 1)
+        { // previously working
+            return -1;
+        }
+        if (status == -1 && dbStatus == -1)
+        { // start working
+            status = 1;
+            // write status on database
+            PGresult *result;
+            std::string order;
+            if (PQstatus(dbconn) == CONNECTION_OK)
+            {
+                order = "Update requests \n SET status = 1;";
+                result = PQexec(dbconn, order.c_str());
+                PQclear(result);
+            }
+            return 1;
+        }
+        if (status && dbStatus)
+        { // working with connection
+            return 1;
+        }
     }
 
     return -1;
@@ -104,46 +113,57 @@ void Database::readAlgorithm()
 
 void Database::saveMESmessage(char *msg, int day, int new_)
 {
-    PGresult *result;
-    std::string order;
-    std::string str(msg);
-    if (PQstatus(dbconn) == CONNECTION_OK)
+    if (status > 0)
     {
-        if(new_)
-            order = "INSERT INTO mesorder (day, message) VALUES ('" + std::to_string(day) + "','" + str + "');";
-        else
-            order = "UPDATE  mesorder SET message = '" + str + "' WHERE day = '" + std::to_string(day) + "' ;" ;
-        result = PQexec(dbconn, order.c_str());
-        PQclear(result);
+        PGresult *result;
+        std::string order;
+        std::string str(msg);
+        if (PQstatus(dbconn) == CONNECTION_OK)
+        {
+            if (new_)
+                order = "INSERT INTO mesorder (day, message) VALUES ('" + std::to_string(day) + "','" + str + "');";
+            else
+                order = "UPDATE  mesorder SET message = '" + str + "' WHERE day = '" + std::to_string(day) + "' ;";
+            result = PQexec(dbconn, order.c_str());
+            PQclear(result);
+        }
     }
 }
 
 char *Database::getMESmessage(int *day)
 {
-    PGresult *result;
-    std::string order;
-
-    if (PQstatus(dbconn) == CONNECTION_OK)
+    if (status > 0)
     {
-        order = "SELECT * FROM mesorder ;";
-        result = PQexec(dbconn, order.c_str());
-        int last = PQntuples(result);
+        PGresult *result;
+        std::string order;
 
-        char *message = PQgetvalue(result, last - 1, 1);
-        (*day) = atoi(PQgetvalue(result, last - 1, 0));
+        if (PQstatus(dbconn) == CONNECTION_OK)
+        {
+            order = "SELECT * FROM mesorder ;";
+            result = PQexec(dbconn, order.c_str());
+            int last = PQntuples(result);
 
-        PQclear(result);
-        return message;
+            char *message = PQgetvalue(result, last - 1, 1);
+            (*day) = atoi(PQgetvalue(result, last - 1, 0));
+
+            PQclear(result);
+            return message;
+        }
     }
     return NULL;
 }
 
-void Database::CleanTable(std::string dbname){
-    PGresult* result;
-    std::string order;
-    if (PQstatus(dbconn) == CONNECTION_OK) {
-        order = "DELETE FROM " + dbname + ";" ;
-        result = PQexec(dbconn, order.c_str());
-        PQclear(result);
+void Database::CleanTable(std::string dbname)
+{
+    if (status > 0)
+    {
+        PGresult *result;
+        std::string order;
+        if (PQstatus(dbconn) == CONNECTION_OK)
+        {
+            order = "DELETE FROM " + dbname + ";";
+            result = PQexec(dbconn, order.c_str());
+            PQclear(result);
+        }
     }
 }
